+++ src/config/babel/babel.config.js
@@ -3,7 +3,6 @@ module.exports = {
     [
       '@babel/preset-env',
       {
-        modules: false,
         targets: {browsers: 'cover 99.5%, last 3 versions, not ie 11'},
       },
     ],
@@ -11,5 +10,6 @@ module.exports = {
   plugins: [
     '@babel/plugin-proposal-throw-expressions',
     '@babel/plugin-transform-arrow-functions',
+    '@babel/plugin-transform-runtime',
   ],
 }