+++ src/config/babel/tests.babel.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  presets: [
+    [
+      {
+        targets: {node: 'current'},
+      },
+    ],
+  ],
+}