+++ src/config/tests/base.jest.config.js
@@ -1,6 +1,8 @@
 module.exports = {
   moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx', 'json', 'node'],
-  testPathIgnorePatterns: ['/dist/', '__tests__/helpers/'],
+  notify: true,
+  testMatch: ['**/__tests__/**/?(*.)+(spec|test).js?(x)', '**/?(*.)+(spec|test).js?(x)'],
+  testPathIgnorePatterns: ['/dist/', '/node_modules/', '/src/', '/__tests__/helpers/'],
   transform: {
     '^.+\\.jsx?$': 'babel-jest',
     '^.+\\.tsx?$': 'ts-jest',